+++ ChangeLog.md
@@ -18,8 +18,19 @@ to browse the changes between the tags.
 
 See docs/process.md for more on how version tagging works.
 
-5.0.8 (in development)
+6.0.0 (in development)
 ----------------------
+- The minimum versions of browser engines supported by emscripten's generated
+  code were bumped, allowing us to remove our internal support for transpilation
+  via babel:
+    MIN_CHROME_VERSION: 74 -> 85
+    MIN_FIREFOX_VERSION: 68 -> 79
+    MIN_SAFARI_VERSION: 12.2 -> 14.1
+  This allows us to assume that features such as mutable-globals, sign-ext, and
+  threading, are universally available.  Disabling these is no longer possible
+  in emscripten.  If you still need to support extremely old browsers, you can
+  manually transpile the output of emscripten (e.g. using babel for JS and
+  binaryen for wasm). (#26677)
 
 5.0.7 - 04/30/26
 ----------------